% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_cbpf_as.R
\name{f_cbpf}
\alias{f_cbpf}
\title{State transition}
\usage{
f_cbpf(x_tt, phi_x, z_beta)
}
\arguments{
\item{x_tt}{state value in t-1 i.e. x_{t-1}}

\item{phi_x}{autoregressive parameter phi}

\item{z_beta}{result of regressor values i.e. z_{t} (matrix) multiplied by
parameters/coefficients (vector)}
}
\value{
\code{T}-dimensional vector of (deterministically computed) state
transitions (conditional means)
}
\description{
Helper function computing the deterministic state transition, or, to put
differently the one-period ahead conditional mean of the latent state
process.
}
