% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_cpp_vech}
\alias{f_cpp_vech}
\title{State transition}
\usage{
f_cpp_vech(x_tt, phi_x, z_add)
}
\arguments{
\item{x_tt}{particle value in t-1 i.e. x_{t-1}; \code{TTx1}-dimensional
vector (double)}

\item{phi_x}{autoregressive parameter (double)}

\item{z_add}{result of regressor values i.e. Z_{1:TT} (matrix) multiplied by
parameters/coefficients (vector) i.e. a matrix (double)}
}
\value{
deterministic state transition (one-period ahead conditional mean)
as a \code{TTx1}-vector
}
\description{
Helper function computing the deterministic state transition, or, to put
differently, the one-period ahead conditional mean of the latent state
process. This version is required for the MCMC part, not the SMC part.
}
\details{
This function is used internally in the PGAS procedure for MCMC sampling of
parameters when subtracting the mean from the state process: it is applied
per state component \code{d=1,...,DD} on a \code{TTx1}-dimensional state
vector where \code{TT} is the number of time periods of all the \code{DD}
components. This is the reason for \code{z_add} to be a vector as it adds
regressor*beta (i.e. regressors matrix time coefficient vector) change for
all \code{t=1,...,T}.
}
