% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generate_data_sim_t.R
\name{generate_data_t}
\alias{generate_data_t}
\title{Generates time series data for various compositional/count models.}
\usage{
generate_data_t(
  nn,
  TT,
  DD,
  true_params,
  x_levels,
  options_include,
  modelling_reg_types
)
}
\arguments{
\item{nn}{current iteration of the cross sectional dimension}

\item{TT}{number of time periods}

\item{DD}{number of shares/fractions (for Dirichlet or Dirichlet-Multinomial)
or the number of categories for a Multinomial distribution}

\item{true_params}{an object of class "\code{trueParams}" which is a list of
true parameter values and meta information such as distribution,
model dimension, the seed under which the true parameters are generated,
the logical indicators that describe which parameters to generate and their
lengths, as provided by the function \code{\link[=new_trueParams]{new_trueParams()}}}

\item{x_levels}{vector of target "means"/"levels" of the states around which
they fluctuate}

\item{options_include}{a named list of three elements:
\itemize{
\item{\code{include_intercept: }}{logical vector of dimension \code{DD}; if
\code{TRUE} include an intercept at the cross sectional unit for component
\code{d}}
\item{\code{include_policy: }}{logical vector of dimension \code{DD}; if
\code{TRUE} include a policy dummy at the cross sectional unit for
component \code{d}}
\item{\code{include_zeros: }}{numeric vector of dimension \code{DD} with
values 1, 2, 3 or 4:
\itemize{
\item{1: }{A dummy pattern that starts at the beginning with zeros and
jumps after half of the overall time period}
\item{2: }{A dummy pattern that starts at the beginning with ones and
plummets to zeros after half of the overall time period}
\item{3: }{A dummy pattern that starts at the beginning with one, then
plummets to zeros after a third of the overall time periods, and then
reverts back to ones for the last third of the time}
\item{4: }{A dummy pattern that starts at the beginning with zeros,
then jumps to ones after a third of the overall time periods, and then
reverts back to zeros for the last third of the time}
}}
}}

\item{modelling_reg_types}{logical vector of dimension 4 where each component
indicates, if \code{TRUE}, that the corresponding regressor type is to be
generated in the following order:
\itemize{
\item{modelling_reg_types, component 1: }{includes z (linear) regressors}
\item{modelling_reg_types, component 2: }{includes u (linear) regressors}
\item{modelling_reg_types, component 3: }{includes z spline regressors}
\item{modelling_reg_types, component 4: }{includes u spline regressors}
}}
}
\value{
a list of two: \code{[[1]]} -> regressors and \code{[[2]]} -> latent
states
}
\description{
The data is a time series of latent state processes of dimension DD as the
multivariate draws can vary along the time dimension. These are used as
parameters of the response/measurement distribution linked/modelled as a
function of regressors (attached to the latent states) where the regressors
and latent states can vary over time.
}
