% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_generate_data.R
\name{generate_x_z}
\alias{generate_x_z}
\title{Simulates regressors and latent states.}
\usage{
generate_x_z(
  phi_x,
  sig_sq_x,
  bet_x,
  x_level,
  x_sd,
  process_log_scale,
  intercept,
  policy_dummy = FALSE,
  zero_pattern = 0,
  drift = FALSE,
  TT
)
}
\arguments{
\item{phi_x}{autoregressive parameter of latent states}

\item{sig_sq_x}{standard deviation of error term in latent state transition
equation}

\item{bet_x}{regressor coefficients/parameters of latent state process}

\item{x_level}{target levels of latent states around which they fluctuate (a
tuning parameter that ensures that states at each multivariate component of
the response fluctuate around a particular level)}

\item{x_sd}{standard deviation allowed for \code{x_level} target (a tuning
parameter that ensures that deviations from the \code{x_level} target are
not too severe)}

\item{process_log_scale}{logical; if \code{TRUE} process is simulated on the
log-scale}

\item{intercept}{logical; if \code{TRUE}, includes an intercept term i.e. a
constant level regressor}

\item{policy_dummy}{logical; if \code{TRUE} includes a dummy that jumps from
zero to one (e.g. a policy or other jump effect to be modelled)}

\item{zero_pattern}{TO-BE-EXPLAINED-LATER}

\item{drift}{TO-BE-EXPLAINED-LATER}

\item{TT}{number of time periods}
}
\value{
a \code{length(beta_x)}x\code{TT} matrix of regressors and a \code{TT}-dimensional vector of latent states
}
\description{
The core function that simulates the regressors and latent states given
certain parameters.
}
