% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_cbpf_as.R
\name{helper_as}
\alias{helper_as}
\title{helps to compute as weights}
\usage{
helper_as(M, x)
}
\arguments{
\item{M}{lhs}

\item{x}{rhs}
}
\value{
some return value needed within the computations of \code{cbpf_as_R()}
}
\description{
helps to compute as weights
}
