% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_pgas.R
\name{monitor_pgas_mcmc}
\alias{monitor_pgas_mcmc}
\title{Monitors PGAS provess in more detail}
\usage{
monitor_pgas_mcmc(
  current,
  total,
  num,
  burn,
  val_true = NULL,
  val_init,
  current_pars,
  num_pars
)
}
\arguments{
\item{current}{current MCMC iteration}

\item{total}{total MCMC iterations}

\item{num}{total number of printed progress messages; if \code{num = 10},
then progress are printed in terms of 10\%, 20\%, 30\%, ...., 100\%}

\item{burn}{burn in period}

\item{val_true}{true parameter values if a the data is simulated}

\item{val_init}{starting values for parameters}

\item{current_pars}{current MCMC parameter draws as a \code{mm x
num_pars}-dimensional matrix}

\item{num_pars}{number of parameters}
}
\value{
print to screen; no return value
}
\description{
Monitors time in terms of percentage of completed (particle) MCMC iterations.
Moreover, initial values, true values (if simulation is performed) and means
of MCMC sampled parameter values are printed to screen. In this sense, the
progress of PGAS is monitored in more detail.
}
