% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_pgas.R
\name{monitor_pgas_states}
\alias{monitor_pgas_states}
\title{PGAS state monitoring}
\usage{
monitor_pgas_states(
  states_drawn,
  states_comp,
  freeze = 1.5,
  current,
  total,
  num_prints
)
}
\arguments{
\item{states_drawn}{current states drawn via SMC (e.g. bootstrap particle
filter)}

\item{states_comp}{states to be compared to; either states from iteration
before or true state values if available (i.e. if data is simulated)}

\item{freeze}{seconds to freeze the plot of}

\item{current}{current MCMC iteration}

\item{total}{total MCMC iterations}

\item{num_prints}{total number of plots e.g. if \code{num_prints = 10},
then 10 comparison/diagnostic plots of the states are printed}
}
\value{
print to screen; no return value
}
\description{
Displaying the sampled state trajectories of the current and the previoius
iterations: allows to see a frozen state trajectory or whether the state
trajectories seem to look reasonable etc. May be used to compare to true
state trajectories if data is simulated.
}
