% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnorm_c}
\alias{mvrnorm_c}
\title{A cpp-wrapper for the R-internal MASS::mvrnorm()-function.}
\usage{
mvrnorm_c(mu, Sigma)
}
\arguments{
\item{mu}{mean vector of the distribution (arma::vec)}

\item{Sigma}{variance covariance matrix of the distribution (arma::mat)}
}
\value{
one draw from the multivariate normal with dimension taken from
mu/Sigma
}
\description{
The wrapper should produce the exact same multivariate normal random numbers
within rcpp/arma code.
}
\details{
Similar to MASS::mvrnorm() the dimension of the result is taken from
\code{mu}/\code{Sigma} (the other arguments of MASS::mvrnorm are taken as
defaults).
}
