% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generate_data_sim_measurements.R
\name{my_rdirichlet}
\alias{my_rdirichlet}
\title{Generates random samples from Dirichlet distribution}
\usage{
my_rdirichlet(alpha)
}
\arguments{
\item{alpha}{a matrix of alpha parameters of a Dirichlet distribution}
}
\value{
a \code{n x D} dimensional matrix of Dirichlet draws
}
\description{
Generates random samples from Dirichlet distribution; the dimension of the
Dirichlet distribution (i.e. the number of shares or fractions) is taken as
the number of columns in the alpha-matrix; the number of samples are taken as
the number of rows of the alpha matrix; hence, each row of \code{alpha}
corresponds to \code{D} alpha parameters for which a \code{D}-dimensional
random draw from the Dirichlet is generated.
}
