% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_A_my_rdistribution.R
\name{my_rmultinomial}
\alias{my_rmultinomial}
\title{Generates random samples from a mulinomial distribution}
\usage{
my_rmultinomial(probs, num_counts)
}
\arguments{
\item{probs}{probabilities for the different categories of the multinomial
distribution given as a matrix}

\item{num_counts}{counts for the different categories of the multinomial}
}
\value{
a nxD dimensional matrix of multinomial draws
}
\description{
Generates random samples from multinomial distribution; the dimension of the
multinomial distribution (i.e. the number of counts that fall into the
categories) is taken as the number of columns in the \code{probs}-matrix; the
number of samples are taken as the rows of the \code{probs} matrix; hence,
each row of \code{robs} corresponds to the \code{D} parameters of the
mulitinomial distribution for which a \code{D}-dimensional random draw is
generated, and all these \code{n} draws are returned in matrix form.
}
