% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pgas_cpp}
\alias{pgas_cpp}
\title{Particle Gibbs with ancestor sampling (PGAS)}
\usage{
pgas_cpp(N, NN, TT, DD, MM, data, Z, priors, par_init, traj_init)
}
\arguments{
\item{N}{number of particles}

\item{NN}{cross sectional dimension}

\item{TT}{time series dimension}

\item{DD}{number of dirichlet fractions/shares i.e. categories}

\item{MM}{PGAS iterations i.e. MCMC iterations (which is equal to the number
of iterations of the SMC-part)}

\item{Z}{regressors contained in the latent state process part}

\item{priors}{hyperpriors for inverted gamma priors of the state process
error variances}

\item{par_init}{initial parameters i.e. starting values for the MCMC part of
the overall PGAS procedure}

\item{traj_init}{initial latent state values i.e. starting values for the
SMC part of the overall PGAS procedure}

\item{y}{measurements: dirichlet fractions/shares}

\item{num_counts}{measurements: dirichlet-multinomial total counts per time
period (\code{T}-dimensional vector)}
}
\value{
List of parameter MCMC samples and latent state trajectory outputs
}
\description{
Runs PGAS with various possible SMC procedures and Gibbs blocks. In this
case we use Armadillo random numbers for the MCMC part and the arma version
of the conditional SMC (in principle, all combinations are possible e.g.
using Rcpp/base-R random numbers for MCMC while the SMC procedure may rely
on arma random numbers)
}
