% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_pgas.R
\name{pgas_out_2_list}
\alias{pgas_out_2_list}
\title{Transforms PGAS output to list format for comparison/testing reasons}
\usage{
pgas_out_2_list(pgas_out, DD, NN, MM)
}
\arguments{
\item{pgas_out}{output of \code{pgas_cpp()} or \code{pgas_R()}}

\item{DD}{multivariate response/measurement dimension: e.g. number of
shares/fractions if the measurements are from a Dirichlet}

\item{NN}{cross sectional dimension}

\item{MM}{number of overall (particle) MCMC iterations}
}
\value{
a list of 19 elements corresponding to former output format
}
\description{
Transforms PGAS output to list format for comparison/testing reasons
}
