% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{propagate_bpf}
\alias{propagate_bpf}
\title{Propagates particles forward}
\usage{
propagate_bpf(mmu, sdd, N)
}
\arguments{
\item{mmu}{mean vector of type arma::colvec}

\item{sdd}{standard deviation as double}

\item{N}{number of particles (int)}
}
\value{
a vector of forward sampled (i.e. propagated) particles of dimension
\code{N}x1 (\code{arma::colvec})
}
\description{
As the bootstrap particle propagates particles forward via the state
transition equation we use the multivariate normal with appropriate mean
vector and standard deviation(in all our models we have a Gaussian state
transition).
}
