% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample}
\alias{resample}
\title{Resampling function}
\usage{
resample(weights, N, id_as_lnspc)
}
\arguments{
\item{weights}{arma::colvec of dimension \code{N} storing particle weights}

\item{N}{number of particles (int)}

\item{id_as_lnspc}{a arma::uvec starting from 1:N; redundant if R::sample()
is used but necessary for the Armadillo functionality}
}
\value{
a arma::uvec of dimension \code{N} containing the resampled indices
}
\description{
Either uses arma random numbers or R's random numbers (via Rcpp::sample) to
do a resampling step i.e. permuting the ancestor indices. Comment in/out the
corresponding parts required.
}
