% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_final_trajectory}
\alias{sample_final_trajectory}
\title{Samples final particle trajectory index}
\usage{
sample_final_trajectory(weights, N, id_as_lnspc)
}
\arguments{
\item{weights}{arma::colvec of dimension \code{N} storing particle weights}

\item{N}{number of particles (int)}

\item{id_as_lnspc}{a arma::uvec starting from 1:N; redundant if R::sample()
is used but necessary for the Armadillo functionality}
}
\value{
returns sampled index (as double; check if int-type could be used)
}
\description{
Last step of conditional SMC/BPF algorithm generates a particle trajectory
to output s.th. PGAS procedure can condition on this draw.
}
