% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_phi_oob}
\alias{test_phi_oob}
\title{helper function to check whether phi-parameter is out of bounds}
\usage{
test_phi_oob(phi, eps)
}
\arguments{
\item{phi}{autoregressive parameter of the state process (double) which is
tested to be smaller than one in absolute value}

\item{eps}{precision (double); how close phi is allowed to be to +1 or -1}
}
\value{
logical; if \code{TRUE}, then \code{phi} is smaller than one in
absolute value given the precision \code{eps}
}
\description{
helper function to check whether phi-parameter is out of bounds
}
