% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_as_c}
\alias{w_as_c}
\title{Computes the ancestor sampling weights.}
\usage{
w_as_c(mean_diff, vcm_diag, log_weights, N, id_as_lnspc)
}
\arguments{
\item{mean_diff}{difference matrix of mean values required (see the formal
derivations of the ancesor weights in the project summary) (arma::mat)}

\item{vcm_diag}{the variance-covariance matrix of the \code{DD}-dimensional
(conditional) state process i.e. the error term variances stacked along
d=1,...,DD (arma::rowvec)}

\item{log_weights}{logarithmic particle weights \code{Nx1}-dimensional
vector (arma::vec); see the derivations of the ancestor weights in the
project summary for details}

\item{N}{number of particles (integer)}

\item{id_as_lnspc}{a arma::uvec starting from 1:N; redundant if R::sample()
is used but necessary for the Armadillo functionality}
}
\value{
ancestor weights of dimension \code{Nx1}-dimensional (for each
particle)
}
\description{
Computes the ancestor sampling weights.
}
