% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w_cbpf}
\alias{w_cbpf}
\title{SMC weights}
\usage{
w_cbpf(N, DD, num_counts, y, xa, id_x)
}
\arguments{
\item{N}{number of particles (int)}

\item{DD}{number of state components (dirichlet fractions or number of
components in the multivariate latent state component) (int)}

\item{num_counts}{number of overall counts per t=1,...,TT (part of the
measurement data) i.e. a scalar int-value for the current time period}

\item{y}{Dirichlet fractions/shares of dimension \code{DD} (part of the
measurement data) observed a specific t=1,...,TT; (arma::rowvec)}

\item{xa}{particle state vector; \code{NxDD}-dimensional arma::vec (as the
whole state vector has \code{DD} components and \code{N} is the number of
particles)}

\item{id_x}{index vector giving the location of the N-dimensional components
for each subcomponent d=1,...,DD within the \code{NxDD} dimensional
\code{xa}}
}
\value{
particle weights
}
\description{
Computes normalized bootrstrap particle weights.
}
\details{
Can currently be used for Dirichlet-multinommial model only.
}
