% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_cbpf_as.R
\name{w_cbpf_R}
\alias{w_cbpf_R}
\title{Computes normalized particle weights}
\usage{
w_cbpf_R(y, N, xa, num_counts, D = DD)
}
\arguments{
\item{y}{measurements: dirichlet-multinomial fractions/shares; matrix of
dimension \code{DDxTT}}

\item{N}{number of particles}

\item{xa}{state trajectories: matrix of DDxTT}

\item{num_counts}{measurement: dirichlet-multinomial total counts per time
period; vector of dimension \code{TT}}

\item{D}{number of categories/classes}
}
\value{
\code{N}-dimensional vector of normalized weights
}
\description{
Computes normalized particle weights for particles in the \code{cbpf_as_R()}
function.
}
