% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_generate_data.R
\name{f}
\alias{f}
\title{State transition}
\usage{
f(x_tt, z, phi_x, bet_x)
}
\arguments{
\item{x_tt}{state value in t-1 i.e. x_{t-1}}

\item{z}{regressor values i.e. z_{t} (matrix)}

\item{phi_x}{autoregressive parameter phi}

\item{bet_x}{regressor parameters/coefficients at z_{t} (matrix)}
}
\value{
\code{T}-dimensional vector of (deterministically computed) state
transitions (conditional means)
}
\description{
Helper function computing the deterministic state transition, or, to put
differently the one-period ahead conditional mean of the latent state
process.
}
