% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_generate_data.R
\name{generate_data_t}
\alias{generate_data_t}
\title{Generates time series data for various models.}
\usage{
generate_data_t(
  distribution = "dirichlet",
  TT,
  DD,
  par_true,
  x_levels,
  x_log_scale,
  intercept_include
)
}
\arguments{
\item{distribution}{specifies the distribution; "dirichlet", "multinomial" or
"dirichlet-multinomial"}

\item{TT}{number of time periods}

\item{DD}{number of shares/fractions (for dirichlet or dirichlet-multinomial)
or the number of categories for a multinomial distribution}

\item{par_true}{list of true parameters that describe the latent state
process}

\item{x_levels}{vector of target "means"/"levels" of the states around which
they fluctuate}

\item{x_log_scale}{logical; if \code{TRUE}, then \code{x_levels[i]} ist taken
as log and the random number generation for the states is performed in logs}

\item{intercept_include}{logical vector; if component is \code{TRUE},
includes an intercept at the cross sectional unit}
}
\value{
a list of two: \code{[[1]]} -> regressors and \code{[[2]]} -> latent
states
}
\description{
The data is a time series, for a given cross sectional unit, from a
Dirichlet, Multinomial or Dirichlet-multinomial distribution. The mulivariate
draws can vary along the time dimension. This is because the parameter of the
distributions that generate the draws are modelled as a function of
regressors and latent states where the regressors and latent states can vary
over time.
}
