% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_generate_data.R
\name{generate_data_t_n}
\alias{generate_data_t_n}
\title{Generates panel data for various models.}
\usage{
generate_data_t_n(
  distribution,
  NN,
  TT,
  DD,
  par_true,
  x_levels,
  x_log_scale,
  include_intercept,
  include_policy,
  include_zeros,
  plot_measurements = FALSE,
  plot_states = FALSE,
  plot_states_each_d = FALSE,
  seed_no = NULL
)
}
\arguments{
\item{distribution}{specifies the distribution; "dirichlet", "multinomial" or
"dirichlet-multinomial"}

\item{NN}{number of cross sectional units}

\item{TT}{number of time periods}

\item{DD}{number of shares/fractions (for dirichlet or dirichlet-multinomial)
or the number of categories for a multinomial distribution}

\item{par_true}{list of true parameters that describe the latent state
process}

\item{x_levels}{target "mean" levels of the states around which they
fluctuate}

\item{x_log_scale}{logical; if \code{TRUE}, x_levels are taken as logs and
the random number generation for the states is performed in logs}

\item{include_intercept}{logical vector of dimension \code{DD}; if \code{TRUE}
include an intercept at the cross sectional unit for component \code{d}}

\item{include_policy}{logical vector of dimension \code{DD}; if \code{TRUE}
include a policy dummy at the cross sectional unit for component \code{d}}

\item{include_zeros}{numeric vector of dimension \code{DD} with values 1,
2, 3 or 4:
\itemize{
\item{1: }{A dummy pattern that starts at the beginning with zeros and
jumps after half of the overall time period}
\item{2: }{A dummy pattern that starts at the beginning with ones and
plummets to zeros after half of the overall time period}
\item{3: }{A dummy pattern that starts at the beginning with one, then
plummets to zeros after a third of the overall time periods, and then
reverts back to ones for the last third of the time}
\item{4: }{A dummy pattern that starts at the beginning with zeros,
then jumps to ones after a third of the overall time periods, and then
reverts back to zeros for the last third of the time}
}}

\item{plot_measurements}{logical; if \code{TRUE}, measurements are plotted
per cross sectional unit \code{n=1,...,N}}

\item{plot_states}{logical; if \code{TRUE}, latent states are plotted
per cross sectional unit \code{n=1,...,N} with a joint plot of all
components together}

\item{plot_states_each_d}{logical; if \code{TRUE}, latent states are
plotted per cross sectional unit \code{n=1,...,N} with a seperate plot for
each component}

\item{seed_no}{integer; random seed to set at the beginning, set \code{NULL}
if not required}
}
\value{
NN-dimensional list of lists of two: \code{[[1]]} -> regressors and
\code{[[2]]} -> latent states
}
\description{
The data is a panel data from a Dirichlet, Multinomial or
Dirichlet-multinomial distribution. The mulivariate draws can vary along the
time and cross-sectional dimensions. This is because the parameter of the
distributions that generate the draws are modelled as a function of
regressors and latent states where the regressors and latent states can vary
over time and cross section.
}
