% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_helper_pgas.R
\name{monitor_pgas_time}
\alias{monitor_pgas_time}
\title{Monitors PGAS process}
\usage{
monitor_pgas_time(current, total, num)
}
\arguments{
\item{current}{current MCMC iteration}

\item{total}{total MCMC iterations}

\item{num}{total number of printed progress messages; if \code{num = 10},
then progress are printed in terms of 10\%, 20\%, 30\%, ...., 100\%}
}
\value{
print to screen; no return value
}
\description{
Monitors time in terms of percentage of completed (particle) MCMC iterations.
In this sense, the progress of PGAS is monitored.
}
