% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_generate_data.R
\name{my_rdirichlet}
\alias{my_rdirichlet}
\title{Generates random samples from dirichlet distribution}
\usage{
my_rdirichlet(alpha)
}
\arguments{
\item{alpha}{alpha parameters of dirichlet distribution given as a matrix}
}
\value{
a nxD dimensional matrix of Dirichlet draws
}
\description{
Generates random samples from dirichlet distribution; the dimension of the
Dirichlet distribution (i.e. the number of shares of fractions) is taken as
the number of columns in the alpha-matrix; the number of samples are taken
as the rows of the alpha matrix; hence, each row of \code{alpha} corresponds
to \code{D} alpha parameters for which a \code{D}-dimensional random draw
is generated, and all these \code{n} draws are returned in matrix form.
}
