% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pgas_R.R
\name{pgas_R}
\alias{pgas_R}
\title{Particle Gibbs with ancestor sampling}
\usage{
pgas_R(N, MM, NN, TT, DD, data, Z, priors, par_init, traj_init, true_states)
}
\arguments{
\item{N}{number of particles}

\item{MM}{number of overall (particle) MCMC iterations}

\item{NN}{cross sectional dimension}

\item{TT}{time series dimension}

\item{DD}{multivariate response/measurement dimension: e.g. number of
shares/fractions if the measurements are from a Dirichlet}

\item{data}{a list of data objects i.e. measurements: e.g. can be dirichlet
fractions and/or number of counts per category (only the latter if
measurements are from a multinomial, and both if measurements come from a
multinomial-dirichlet)}

\item{Z}{regressor matrix of Z_{t}'s}

\item{priors}{inverteg gamma prior (hyyper-)parameters}

\item{par_init}{initial values of parameters}

\item{traj_init}{initial state trajectory}

\item{true_states}{true laten states passed from simulated data for testing
purposes}
}
\value{
a list with components being: all MCMC parameter draws and all drawn
state trajectories (smc outuput)
}
\description{
R version: MCMC parameters are drawn with R functions, SMC-part (e.g.
bootstrap particle filter) can be a C++ or R function (copy paste or comment
in/out)
}
